categoryId가 0인 조건을 isAllCategories 메소드명으로 명시

+++ src/main/java/com/nts/reservation/product/dao/ProductDao.java
@@ -32,15 +32,15 @@ public class ProductDao {
 		this.jdbc = new NamedParameterJdbcTemplate(dataSource);
 	}
 
-	public List<Product> selectProducts(long start, long limit) {
-		Map<String, Long> params = new HashMap<>();
+	public List<Product> selectProducts(int start, int limit) {
+		Map<String, Integer> params = new HashMap<>();
 		params.put("start", start);
 		params.put("limit", limit);
 		return jdbc.query(SELECT_PRODUCT, params, rowMapper);
 	}
 
-	public List<Product> selectProductsByCategory(long categoryId, long start, long limit) {
-		Map<String, Long> params = new HashMap<>();
+	public List<Product> selectProductsByCategory(int categoryId, int start, int limit) {
+		Map<String, Integer> params = new HashMap<>();
 		params.put("categoryId", categoryId);
 		params.put("start", start);
 		params.put("limit", limit);
@@ -51,8 +51,8 @@ public class ProductDao {
 		return jdbc.queryForObject(SELECT_PRODUCT_COUNT, Collections.emptyMap(), Integer.class);
 	}
 
-	public int selectProductsCountByCategory(long categoryId) {
-		Map<String, Long> params = new HashMap<>();
+	public int selectProductsCountByCategory(int categoryId) {
+		Map<String, Integer> params = new HashMap<>();
 		params.put("categoryId", categoryId);
 		return jdbc.queryForObject(SELECT_PRODUCT_COUNT_BY_CATEGORY, params, Integer.class);
 	}
